Rename clean queries file
